Water Consumption Checking Data

+++ NEMBlockchain/NEMBlockchain.Contract/Water/WaterConsumtionTotalContract.cs
@@ -6,7 +6,7 @@ namespace NEMBlockchain.Contract.Water
     public class WaterConsumtionTotalContract
     {
         public string UserId { get; set; }
-        public double? TotalVolume { get; set; }
+        public decimal? TotalVolume { get; set; }
         public DateTime LogTime { get; set; }
         public bool isExistedOnNem { get; set; }
     }
