+++ NEMBlockchain/NEMBlockchain.Data/AutoFlowDB_Water_Water_DataContext/ViewModels/WaterConsumptionDetailViewModel.cs
@@ -8,7 +8,7 @@ namespace NEMBlockchain.Data.AutoFlowDB_Water_Water_DataContext.ViewModels
     {
         public string UserId { get; set; }
         public string FunitureName { get; set; }
-        public double? Volume { get; set; }
+        public decimal? Volume { get; set; }
         public DateTime? LogTime { get; set; }
     }
 }
