+++ NEMBlockchain/NEMBlockchain.Data/AutoFlowDB_Water_Water_DataContext/ViewModels/WaterConsumtionTotalViewModel.cs
@@ -7,7 +7,7 @@ namespace NEMBlockchain.Data.AutoFlowDB_Water_Water_DataContext.ViewModels
     public class WaterConsumtionTotalViewModel
     {
         public string UserId { get; set; }
-        public double? TotalVolume { get; set; }
+        public decimal? TotalVolume { get; set; }
         public DateTime LogTime { get; set; }
     }
 }
